tvdemo: fix 32-bit/64-bit streaming incompatibility in TAsciiChart

+++ examples/tvdemo/ascii.h
@@ -70,7 +70,7 @@ public:
 
 private:
 
-    long asciiChar;
+    uchar asciiChar;
 
     virtual const char *streamableName() const
          { return name; }
